Tests run quicker in parallel
Rather than running tests sequentially this commit switches to parallel testing. This means we can drop the requirement for existing resources (so no more _ID env vars required) but also means we need to build our dependencies for each test. In some cases we need to amend resource names so Cosmic does not get upset when a resource with that name alreay exists.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>