Store offering type as case insensitive value
A bug exists where terraform tries to update a resource's offering type
if the case supplied by user does not match what's in the state. This
causes an error from the Cosmic API as it returns a message that the
offering type is already set.

This was found in the instance offering but the fix has been extended to
other resources with an offering field.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>